+++ src/arpes/correction/background.py
@@ -13,7 +13,7 @@ from arpes.utilities import normalize_to_spectrum
 __all__ = ("remove_incoherent_background",)
 
 LOGLEVELS = (DEBUG, INFO)
-LOGLEVEL = LOGLEVELS[0]
+LOGLEVEL = LOGLEVELS[1]
 logger = getLogger(__name__)
 fmt = "%(asctime)s %(levelname)s %(name)s :%(message)s"
 formatter = Formatter(fmt)
