+++ src/arpes/correction/fermi_edge.py
